resume order after 20% video hits/plays
Before this change, if the 20% chance hits, it would play silent night, then i would be set back to 1. This fix plays silent night, then sets i back to whatever it was before (as it was already incremented after the previous video played), so the sequence continues on the next tap.